Crate pallet_account_linker

AccountLinker Pallet

The AccountLinker pallet provides functionality for linking a Litentry account to account at other networks. (currently support Ethereum (BSC), BTC and Polkadot ecosystem)

Overview

The AccountLinker pallet stores the linking relation between Litentry accounts and accounts at other networks. It also offers extrinscs for user to update the linking relation. For each linking relation, user may choose to freshly link new account or replace an existing linked account with a new provided one.

Interface

Dispatchable Functions

link_eth - Link an Ethereum address to a Litentry account providing a proof signature from the private key of that Ethereum address.

link_btc - Link an BTC address to a Litentry account providing a proof signature from the private key of that BTC address.

link_polkadot - Initiate a link request to link a Litentry address to another Litentry address

accept_polkadot - Accept a pending link_polkadot request to link a Litentry address to another Litentry address.

Re-exports

pub use pallet::*;

Modules

pallet
The module that hosts all the FRAME types needed to add this pallet to a runtime.

weights
Autogenerated weights for pallet_account_linker